backend/pdf_ingestion/requirements.txt
fastapi
uvicorn
langchain
langchain-pymupdf4llm
langchain-openai
langchain-community
python-multipart     
httpx                
python-dotenv